Fix bug with are_states_equal function by removing unnecessary if blocks
Summary:
Remove unnecessary if blocks contained in `are_states_equal` function. Currently will trigger a bug if run on a `OptimizerModule` or `COMPATIBLE_DATA_STRUCTURES`. Since this function is designed for testing whether two nested state dictionaries are equal, we can simply remove those if blocks to prevent this error.

Does not impact any functionality with `OptimizerModule` - it is purely for debugging.

Reviewed By: kaushik88

Differential Revision: D44157758

fbshipit-source-id: 7434c519822555f06710c15121752bf847be8fd1