Fix checkpointing bug caused by inconsistent state[EXP_AVG] updating in Distributed Shampoo
Summary:
In the process of enabling `torch.distributed.checkpoint` for external checkpointing, discovered that we are not updating `state[EXP_AVG]` consistently across all devices even though it is allocated as a Tensor. This breaks the stateful protocol and means that we are loading a zeros tensor from checkpoint for some of the parameters!

This diff corrects this by removing the additional if condition which prevents `EXP_AVG` from being updated across all parameters.

Reviewed By: shintaro-iwasaki

Differential Revision: D46755200

fbshipit-source-id: eec12d4a04686bdcfe8545f57d80b4e232545afb