Raise ValueError instead of warning for observing inf/nan in factor matrices in Shampoo
Summary:
Currently, we only throw an error with Distributed Shampoo when one observes nan or inf values within the root inverse matrix. We raise a warning when we observe nan or inf in the factor matrix.

However, this is not informative when either data corruption or weird behavior from the trainer impacts Shampoo's states. It is also guaranteed that when inf/nan appears in the factor matrix, the eigendecomposition solver will also admit inf and nan values.

Therefore, we choose to fail earlier. We also add additional logging in order to help engineers mitigate the issue.

Differential Revision: D51734920

fbshipit-source-id: 100acdbd7759ffaa4d7a524d5a55bf6eb0eea703