Add AlgoPerf improvements into OSS PyTorch Shampoo
Summary:
We provide consistent or similar modifications to the Shampoo code as used in the PyTorch Shampoo AlgoPerf submission.

This incorporates two additional features:
1. Support for an additional EMA factor (`beta3`).

We add a new field `beta3` used for computing the exponential moving average of the gradient only used for the current iteration. `beta1` is solely used for updating the EMA of the gradient stored as the optimizer state for subsequent iterations. This is similar to how multiple EMAs of the gradient is used in the Lion algorithm (see https://arxiv.org/abs/2302.06675).

Mathematically, ignoring momentum, given a preconditioner $A_k$ and learning rate $\alpha_k > 0$, we are performing the updates
$$\bar{g}_k = \beta_1 \bar{g}_{k - 1} + (1 - \beta_1) g_k,$$
$$w_{k + 1} = w_k - \alpha_k A_k^{-1/2} (\beta_3 \bar{g}_{k - 1} + (1 - \beta_3) g_k).$$
If $\beta_1 = \beta_3$, then this is equivalent to using a single EMA factor.

This enables us to reproduce the changes used by NAdamW from the AlgoPerf contest, where the current iteration applies the same EMA term a second time, i.e.,
$$\bar{g}_k = \beta_1 \bar{g}_{k - 1} + (1 - \beta_1) g_k$$
$$w_{k + 1} = w_k - \alpha_k A_k^{-1/2} (\beta1 \bar{g}_k + (1 - \beta_1) g_k)$$

Note that this is equivalent to first set of equations by setting $\beta3 = \beta_1^2$.

One subtle change we are adding in addition to this is to modify the bias correction term. We did not handle this properly within our AlgoPerf submission. Namely, instead of using a bias correction of $1 - \beta_1^k$, we should instead use a bias correction of $1 - \beta_3 \cdot \beta_1^{k - 1}$.

As observed in the AlgoPerf submission, using a smaller EMA factor at the current iteration appears to yield better performance. This is also consistent with what Lion has discovered (in conjunction with sign SGD). We notably apply this to both Shampoo and the grafting updates (not solely for grafting).

2. Support for a dampening factor within momentum (similar to PyTorch SGD's implementation).

While we had previously supported momentum and Nesterov (applied to the search directions) through the form:
$$m_k = \eta m_{k - 1} + p_k$$
$$w_{k + 1} = w_k - \alpha_k m_k,$$
we are now incorporating a dampening term $\bar{\eta}$ of the form:
$$m_k = \eta m_{k - 1} + (1 - \bar{\eta}) p_k$$
$$w_{k + 1} = w_k - \alpha_k m_k.$$

This turns out to be beneficial for tuning momentum independently from the learning rate, specifically when the dampening factor is set equal to the momentum parameter. A similar technique is used for Nesterov applied to the search directions.

Reviewed By: tsunghsienlee

Differential Revision: D60572125

fbshipit-source-id: 96459b58cbee1cf4de44bf9377db6f6f36996297