Fix `with_comms` usage
Summary: The introduction of `with_comms` as a decorator with an optional argument in https://github.com/pytorch/pytorch/pull/138108 is not backwardly compatible with existing usages of `with_comms`. The correct usage is now `with_comms()`. However, the current CI testing environment is using PyTorch 2.5.1, which does not include this change. Therefore, the with_comms usage should be updated to match the PyTorch 2.5.1 version

Reviewed By: anana10c

Differential Revision: D65426664

fbshipit-source-id: 97b6fbefe12ebf01ef0d77fb972d5b5a4afd4922