Avoid using local modules in some unittests (#49)
Summary:
To avoid installation issues like https://github.com/facebookresearch/optimizers/issues/46, I added the Python `-I` flag when running two of the unittests.

Note that the unittests for which I didn't add this flag import from the `distributed_shampoo.tests` directory that is not installed as part of the package, so trying to run these tests with the `-I` flag will lead to an import error.

Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/49

Reviewed By: anana10c

Differential Revision: D65958371

Pulled By: tsunghsienlee

fbshipit-source-id: cc861ceff6779ea83059f50701a7a3357dab4e4c