Add workflow that runs examples (#53)
Summary:
Added a workflow that runs the default and DDP example on CPU and GPU for one epoch. ~This workflow only runs when a PR is created and not on every push, in contrast to all other workloads.~ *Edit: The workflow runs on `push` and `pull_request`, like all other workflows.*

I think the FSDP, HSDP, and fully shard examples cannot be run on a single GPU since the PyTorch source code sets `sharding_strategy=NO_SHARD` when the world size is 1. Because of this the parameters are not flattened, which will lead to an error [here](https://github.com/facebookresearch/optimizers/blob/main/distributed_shampoo/utils/shampoo_fsdp_distributor.py#L358-L361).

In the future it would be great to make this a proper regression test and 1) check that the loss value after one epoch stays the same and 2) that the update step time stays the same (both within some threshold).

Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/53

Reviewed By: anana10c

Differential Revision: D66139971

Pulled By: tsunghsienlee

fbshipit-source-id: b2f964ebca538cc8e76fde8b0da3e3c392c9a18b