Simplify the input argument of `_construct_global_block_info_list` (#70)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/70

Instead of passing the entire `buffer_size_ranks` tuple, the `group_source_ranks` tuple is extracted from it and passed as an argument. This simplifies the code and makes it more readable. Furthermore, this might help the consolidation of `_construct_global_block_info_list` and `_construct_local_block_info_list` in the future.

Reviewed By: anana10c

Differential Revision: D67606282

fbshipit-source-id: 7aefea61be746f8ae967d0e21bc32708e52b20c1