Refactor AdamGraftingConfig to inherit from RMSPropGraftingConfig (#74)
Summary:

This diff refactors `AdamGraftingConfig` to inherit from `RMSPropGraftingConfig`, taking advantage of the fact that `Adam` is essentially `RMSProp` with momentum. This inheritance was previously blocked by an `isinstance` check, but is now enabled by the switch to `type` checking.

Additionally, this diff fixes an oversight in [the previous `type.__subclasses__()` refactor](https://github.com/facebookresearch/optimizers/pull/72), ensuring that parent classes are properly included.

Reviewed By: anana10c

Differential Revision: D67719358