Refactor AdamGraftingConfig to inherit from RMSPropGraftingConfig (#74)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/74

This diff refactors `AdamGraftingConfig` to inherit from `RMSPropGraftingConfig`, taking advantage of the fact that `Adam` is essentially `RMSProp` with momentum. This inheritance was previously blocked by an `isinstance` check, but is now enabled by the switch to `type` checking.

Additionally, this diff fixes an oversight in [the previous `type.__subclasses__()` refactor](https://github.com/facebookresearch/optimizers/pull/72), ensuring that parent classes are properly included.

Reviewed By: anana10c

Differential Revision: D67719358

fbshipit-source-id: 24a0264bd3d314f2b5b370af3b075dde4773fb4f