Fix incorrectly PreconditionerConfig instantiation in eigendecomposed Shampoo tests (#100)
Summary:

This diff fixes an issue in the eigendecomposed Shampoo tests where the PreconditionerConfig was incorrectly instantiated. The code changes include adding the amortized computation properties to the PreconditionerConfig and updating the inverse exponent override.

This bug was hided due to our test setup is weak and the existence of `is_factor_matrice_diagonal`; as a result, even with incorrect setting, matrix functions still generate the preferred output even though the setup is totally incorrect.

Also fix some miscalculations in comment.

Differential Revision: D71863442