Fix incorrectly PreconditionerConfig instantiation in eigendecomposed Shampoo tests (#100)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/100

This diff fixes an issue in the eigendecomposed Shampoo tests where the PreconditionerConfig was incorrectly instantiated. The code changes include adding the amortized computation properties to the PreconditionerConfig and updating the inverse exponent override.

This bug was hidden due to our test setup is weak and the existence of `is_factor_matrice_diagonal`; as a result, even with incorrect setting, matrix functions still generate the preferred output even though the setup is totally incorrect.

Also fix one miscalculation in comment.

Reviewed By: anana10c

Differential Revision: D71863442

fbshipit-source-id: fde9bf743525b842ee6654536f7f0df5f758c956