Leverage enum.unique enforce the uniqueness property
Summary: This diff introduces the use of the [`enum.unique`]((https://docs.python.org/3/library/enum.html#enum.unique)) decorator in several files to enforce the uniqueness property of the enum classes. This ensures that the values of the enum classes are unique and prevents any potential conflicts or errors that may arise from duplicate values.

Reviewed By: minddrummer

Differential Revision: D72679390

fbshipit-source-id: f47c05bf9a073350d2b74db0739387316c8034cd