Revert the usage of `instantiate_device_type_tests()`
Summary: The current `device` argument provided from `instantiate_device_type_tests()` under the setting of `ShampooDDPDistributorDeviceTest` is always `"cuda: 0"` without considering the ranks. This is an incorrect setting in multiple devices. As a result, revert the usages of `instantiate_device_type_tests()` to make sure the Nvidia GPU and AMD GPU tests back to normal.

Reviewed By: hjmshi

Differential Revision: D73214634

fbshipit-source-id: 96bc8f86ec1088241c7f1a70a934da45420871da