Add `abs_epsilon` into `CoupledHigherOrderConfig`
Summary: This diff adds an `abs_epsilon` parameter to the `CoupledHigherOrderConfig` class in the `matrix_functions.py` file. This diff also enforces the constraints the users have to specify both `rel_epsilon` and `abs_epsilon` in order to use higher-order method to prevent oversights on missing the configurations of this usages.

Reviewed By: gajjanag

Differential Revision: D72988561

fbshipit-source-id: 20125184f3d36d8e6860d9aecd061199cbf8c974