Add unit tests for matrix inverse root computation using larger epsilon (#161)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/161

Done by gavinzhang.

This diff added a test case where epsilon is deliberately set to be very large (eps = 80). Manually computed the expected preconditioned gradients to compare with the outputs of shampoo preconditioner list. The goal of this unit test is to ensure the accuracy of the matrix inverse function for larger values of epsilon, so that in the future we might set an algorithmic epsilon that may be deliberately large.

This diff also fixes other bugs related with pseudo-inverse.

Differential Revision: D74108446

fbshipit-source-id: 62599461136ef0442f82061eedf79a0e6b891097