Support distributed shampoo examples for internal runs (#188)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/188

Make distributed shampoo examples work internally on local devservers and MAST, by BUCK-ifying the OSS-only example files and import path conversion.

After this, the examples can run with `torchx`. See `//fbcode/scripts/haoyuz/distributed_shampoo/run_example.sh` for the detailed instructions.

Note: To make the examples easier to run on MAST (which by default does not allow internet access), the package packs a downloaded copy of Cifar-10 dataset. The trainier utils detect and make a copy of the packaged dataset to the designated path instead of downloading it directly when running on MAST.

Reviewed By: tsunghsienlee

Differential Revision: D76478106

fbshipit-source-id: 1c48543da47032f88f585257d8d9ec1fe658287c