Add test on `_eigh_eigenvalue_decomposition()` when retry for double precision
Summary: This diff addresses a bug discovered by anana10c, where the `_eigh_eigenvalue_decomposition()` function would retry with double precision but fail to convert the result back to the original data type. To ensure the correctness of this function, a new test has been added to verify that the result is properly converted back to the original `dtype`.

Reviewed By: anana10c

Differential Revision: D76205885

fbshipit-source-id: 03705f7fe4400f692072e1c35c44fd6a77d22604