Enhance Shampoo on empty tensor
Summary:
The definition of an empty tensor is one with a size containing `torch.Size([0])`. This diff modifies the `merge_small_dims()` function to accommodate empty tensors.

The corresponding check on empty gradients are changed as well in order to accomdate this.

Note that even though the changes are made to `FSDPDistributor` (`shampoo_fsdp_distributor.py`) and `HSDPDistributor` (`shampoo_hsdp_distributor.py`), it is not feasible to test how it interact with empty tensor on a multi-GPUs setting because it won't be able to shard an empty tensor. A potential next step is testing empty tensors on FSDP and HSDP with single GPU (i.e., one shard) setting.

None of our core codes need to be modified becaose those are already well-equipped for empty tensors already.

Reviewed By: runame, gajjanag

Differential Revision: D76104842

fbshipit-source-id: 3ba55a2d81b9967d9036b951aeef727ba36c40d6