Consolidate different `train_model()` usages into the one under `trainer_utils.py`
Summary: This diff consolidates different `train_model()` usages into the one under `trainer_utils.py`. The changes include refactoring the `train_model()` function to be more modular and reusable, as well as updating the code in the example files to use the new function. Additionally, the diff includes changes to the `distributed_shampoo` module to support distributed checkpointing.

Differential Revision: D77452646