Enable adaptive eigendecomposition computation frequency for eigh
Summary:
This diff enables setting a tolerance to control the error induced by infrequent eigendecompositions for `EigendecomposedShampooPreconditionerList` and `EigenvalueCorrectedShampooPreconditionerList`, which adaptively determines the eigendecomposition computation frequency. See section 4 in [this paper](https://www.arxiv.org/abs/2506.03595) for more details.

This is an alternative implementation to [this GitHub PR](https://github.com/facebookresearch/optimizers/pull/177). There is one functional difference to this implementation: when an eigendecomposition is skipped and the previous eigenbasis is reused, the implementation in this diff will use the estimated eigenvalues based on the previous eigenbasis and the current factor matrices, whereas the implementation in the GitHub PR will just reuse the previous eigenvalues. The implementation here is consistent with how this is handled in the QR algorithm. We don't have any empirical results for how the two options influence the convergence behavior.

**Comments:**
- The default tolerance for the QR algorithm has changed to 0.0, which just means always running until the maximum number of iterations is hit.
- I set `repr=False, compare=False` for `EigendecompositionConfig.eigenvectors_estimate`, mostly to avoid errors when trying to print/compare configs where it has not been set (yet). This could hypothetically be dangerous if users want to check two configs for equality and include `eigenvectors_estimate` in the comparison. However, the user should never touch the configs after instantiation of `DistributedShampoo` and `eigenvectors_estimate` will not be initialized before that, so I think it should be safe to exclude it from comparisons.

Reviewed By: tsunghsienlee

Differential Revision: D76645710

fbshipit-source-id: 6c1950f24d96a5333adead0d4269d8f3f6598a59