Consolidate different `train_model()` usages into the one under `trainer_utils.py` (#196)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/196

This diff consolidates different `train_model()` usages into the one under `trainer_utils.py`. The changes include refactoring the `train_model()` function to be more modular and reusable, as well as updating the code in the example files to use the new function. Additionally, the diff includes changes to the `distributed_shampoo` module to support distributed checkpointing.

Reviewed By: runame

Differential Revision: D77452646

fbshipit-source-id: 0b5b8214de3a783ac2bf647c2260f22e26416fa0