Initialize optimizer state outside of preconditioner lists (#204)
Summary:

This moves the responsibility for initializing the optimizer state away from the preconditioner lists to the optimizer itself by initializing the state as soon as possible. Now the optimizer state does not have to be passed to preconditioner lists that do not have any additional state (and therefore don't have to modify the optimizer state).

Reviewed By: tsunghsienlee

Differential Revision: D78194753