Add SpectralDescentPreconditionerList (#205)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/205

Adding a `SpectralDescentPreconditionerList` (which can only be used for 2D gradients, or gradients that can be reshaped to 2D) with a corresponding `SpectralDescentPreconditionerConfig`.

For example, this can be used to implement Muon by using `SpectralDescentPreconditionerConfig` with
```
orthogonalization_config=NewtonSchulzOrthogonalizationConfig(
    scale_by_dims_fn=lambda d_in, d_out: max(1, d_out / d_in)**0.5,
)
````
for the hidden layers' weight matrices and AdamW for all other parameters.

Reviewed By: tsunghsienlee, hjmshi

Differential Revision: D78026187

fbshipit-source-id: 1d74d814bd3f876c6e245aabde734b1ea5eb5ef5