Remove `self._device` in `DistributedShampoo` class
Summary: There is no need to create a member variable `self._device` when it is used once only. Moreover, this diff changes the access of it from the first element of `MASKED_BLOCKED_PARAMS` to `local_block_params` when the latter one is more accessible while the former one is only consistently accessible during the init time.

Reviewed By: anana10c

Differential Revision: D78501117

fbshipit-source-id: f33f6370e4c993366345453b5f87436964843388