Refactor `_get_function_args_from_config()`
Summary: The previous `_get_function_args_from_config()` implementation has an implicit requirement that `func` is required twice, one for the input argument, and for the function invokes after it. This implicit requirement should not be needed if we decorate the function's arguments initially. As a result, refactor the function to remove this implicit requirement and make it less error-prone to use.

Reviewed By: wz337

Differential Revision: D63863242

fbshipit-source-id: b54deb6b1a004e9ec16efcfc6b8ecdc3b958a3c0