Replace lambda function in `matrix_function_types.py` and `shampoo_types.py` to a private static method for pickable concern
Summary:
This is the same spirit of D82836543, which changes the default value of `scale_fn` of `SignDescentPreconditionerConfig` to a private static method for pickling concern.

We found a few other occurrences of `lambda` function in `shampoo_types.py` and `matrix_function_types.py`. In order for shampoo state_dict (in particular, `param_groups`) to be compatibale with pytorch, we need to replace all lambda functions with a normal function, because `torch.save()` uses pickle for serialization, and pickle cannot serialize lambda function but only regular function.

Reviewed By: anana10c, hjmshi

Differential Revision: D85902488

fbshipit-source-id: ead3637fe2202ed10681312bf0a0652036be32a9