Sync KL-Shampoo to OSS (#255)
Summary:
Pull Request resolved: https://github.com/facebookresearch/optimizers/pull/255

Sync KL-Shampoo implementation to OSS. Also, add KL-Shampoo reference to README.

**Original summary:**

Adding a simplified version of [KL-Shampoo](https://arxiv.org/abs/2509.03378). The main difference to the algorithm presented in the paper is that here only the update of the factor matrices is adjusted, but the eigenvalue correction to the individual factor matrices is not implemented. This correction can also be used for regular Shampoo and will be added in the future.

KL-Shampoo is implemented by storing either the inverse root matrices directly, or storing the eigendecompositions of the factor matrices, analogous to regular Shampoo.

Reviewed By: anana10c, hjmshi

Differential Revision: D89193149

fbshipit-source-id: 98b0499aba64ec1b4f37e6989941eea9c97d3e6d